Fix the connection on Ubuntu 20.04
I'm not sure if this was a firmware update of the laptop or the upgrade
from Ubuntu 19.10, but once I reset my pairing I wasn't able to pair
anymore.

bluetoothctl would report when trying to pair:
Failed to pair: org.bluez.Error.AuthenticationFailed

Doing it the same way as the hid_keyboard example fixed it, I guess it's
fine with a small enough device name.